requirements.txt
beautifulsoup4==4.3.2
cffi==0.8.6
cryptography==0.7.2
cssselect==0.9.1
decorator==4.0.10
Django==1.7.5
django-hitcount==1.2.2
django-taggit==0.21.2
matplotlib==1.4.3
numpy==1.9.1
oauthlib==0.7.2
pattern==2.6
Pillow==3.0.0
praw==3.5.0
pyasn1==0.1.7
pycparser==2.10
pyHook==1.5.1
pyOpenSSL==0.14
pyparsing==2.0.3
python-dateutil==2.4.0
pytz==2014.10
pywin32==219
requests==2.11.1
requests-oauthlib==0.4.2
six==1.10.0
tweepy==3.3.0
update-checker==0.11
virtualenv==15.0.3
zope.interface==4.1.2
psycopg2==2.5.3